- add `ovos_utils.parse.is_compatible_dict` - add `ovos_utils.parse.delete_key_from_dict` - add `OVOSSkill` template to `ovos_utils.skill.templates.ovos` - add new module `ovos_utils.skill.decorators` - refactor, move `killable_intents` methods to `ovos_utils.skill.decorators` - refactor, move `killable_intents` patches to `OVOSSkill` - refactor, always use `ovos_utils.ensure_mycroft_import` - deprecate `ovos_utils.misc` - deprecate `ovos_utils.parse.MatchStrategy.QUICK_LEV_RATIO`  (invalid option, shouldn't be there) - move `["phoneme_guesser", "colour"]` to `extras_require` in setup.py
