+++ ovos_utils/skills/templates/__init__.py
@@ -1,0 +1,1 @@
+from ovos_utils.skills.templates.ovos import OVOSSkill, OVOSFallbackSkill
