+++ ovos_utils/waiting_for_mycroft/skill_gui.py
@@ -16,24 +16,13 @@ from enum import IntEnum
 from os.path import join, isfile
 from ovos_utils import resolve_ovos_resource_file
 from ovos_utils.waiting_for_mycroft.settings_gui_generator import SettingsGuiGenerator
+from ovos_utils import ensure_mycroft_import
 
-try:
-    from mycroft.enclosure.gui import SkillGUI as _SkillGUI
-    from mycroft.util import resolve_resource_file
-    from mycroft.messagebus.message import Message
-except ImportError:
-    import sys
-    from ovos_utils.log import LOG
-    from ovos_utils import get_mycroft_root
-    MYCROFT_ROOT_PATH = get_mycroft_root()
-    if MYCROFT_ROOT_PATH is not None:
-        sys.path.append(MYCROFT_ROOT_PATH)
-        from mycroft.enclosure.gui import SkillGUI as _SkillGUI
-        from mycroft.util import resolve_resource_file
-        from mycroft.messagebus.message import Message
-    else:
-        LOG.error("Could not find mycroft root path")
-        raise ImportError
+ensure_mycroft_import()
+
+from mycroft.enclosure.gui import SkillGUI as _SkillGUI
+from mycroft.util import resolve_resource_file
+from mycroft.messagebus.message import Message
 
 # implements the following GUI functionality
 # https://github.com/MycroftAI/mycroft-core/pull/2683
