+++ setup.py
@@ -5,7 +5,6 @@ setup(
     version='0.0.7',
     packages=['ovos_utils',
               'ovos_utils.waiting_for_mycroft',
-              'ovos_utils.misc',
               'ovos_utils.intents',
               'ovos_utils.sound',
               "ovos_utils.enclosure",
@@ -13,19 +12,22 @@ setup(
               'ovos_utils.enclosure.mark1.eyes',
               'ovos_utils.enclosure.mark1.faceplate',
               'ovos_utils.skills',
-              'ovos_utils.plugins',
               'ovos_utils.skills.templates',
+              'ovos_utils.skills.decorators',
+              'ovos_utils.plugins',
               'ovos_utils.lang'],
     url='https://github.com/OpenVoiceOS/ovos_utils',
-    install_requires=["phoneme_guesser",
-                      "mycroft-messagebus-client",
-                      "inflection",
-                      "colour",
-                      "pexpect",
-                      "PyYAML",
-                      "kthread",
-                      "json_database",
-                      "requests"],
+    install_requires=[
+        "mycroft-messagebus-client",
+        "pexpect",
+        "PyYAML",
+        "kthread",
+        "json_database",
+        "requests",
+        "inflection"],
+    extras_require={
+        "extras": ["phoneme_guesser", "colour"]
+    },
     include_package_data=True,
     license='Apache',
     author='jarbasAI',
